Move Standard Library and Jet Registry from Nock
Moves standard library and Jet Registry information outside the Nock
module into own separate modules.
