+++ apps/anoma_lib/lib/examples/enock.ex
@@ -44,7 +44,7 @@ defmodule Examples.ENock do
   def counter_logic() do
     [
       counter_arm(),
-      0 | Nock.logics_core()
+      0 | Nock.Lib.logics_core()
     ]
   end
 
@@ -188,7 +188,7 @@ defmodule Examples.ENock do
   @spec dec() :: Noun.t()
   def dec() do
     sample = 999
-    core = [dec_arm(), sample | Nock.logics_core()]
+    core = [dec_arm(), sample | Nock.Lib.logics_core()]
 
     assert Nock.nock(core, [9, 2, 0 | 1]) |> elem(1) |> Noun.equal?(998)
 
@@ -222,7 +222,7 @@ defmodule Examples.ENock do
   @spec cue() :: Noun.t()
   def cue() do
     sample = 999
-    core = [cue_arm(), sample | Nock.logics_core()]
+    core = [cue_arm(), sample | Nock.Lib.logics_core()]
 
     core
   end
@@ -245,7 +245,7 @@ defmodule Examples.ENock do
   @spec jam() :: Noun.t()
   def jam() do
     sample = 999
-    core = [jam_arm(), sample | Nock.logics_core()]
+    core = [jam_arm(), sample | Nock.Lib.logics_core()]
 
     core
   end
@@ -269,7 +269,7 @@ defmodule Examples.ENock do
   @spec sign() :: Noun.t()
   def sign() do
     sample = [999 | 888]
-    core = [sign_arm(), sample | Nock.logics_core()]
+    core = [sign_arm(), sample | Nock.Lib.logics_core()]
 
     valid_args = [ECrypto.blood_msg() | ECrypto.londo().internal.sign]
     invalid_args = [ECrypto.blood_msg() | ECrypto.londo().external.sign]
@@ -304,7 +304,7 @@ defmodule Examples.ENock do
   @spec verify() :: Noun.t()
   def verify() do
     sample = [999 | 888]
-    core = [verify_arm(), sample | Nock.logics_core()]
+    core = [verify_arm(), sample | Nock.Lib.logics_core()]
 
     valid_args = [ECrypto.blood_l_signed() | ECrypto.londo().external.sign]
     invalid_args = [ECrypto.blood_msg() | ECrypto.londo().internal.sign]
@@ -340,7 +340,7 @@ defmodule Examples.ENock do
   @spec sign_detatched() :: Noun.t()
   def sign_detatched() do
     sample = [999 | 888]
-    core = [sign_detatched_arm(), sample | Nock.logics_core()]
+    core = [sign_detatched_arm(), sample | Nock.Lib.logics_core()]
 
     valid_args = [ECrypto.blood_msg() | ECrypto.londo().internal.sign]
 
@@ -373,7 +373,7 @@ defmodule Examples.ENock do
   @spec verify_detatched() :: Noun.t()
   def verify_detatched() do
     sample = [999 | 888]
-    core = [verify_detatched_arm(), sample | Nock.logics_core()]
+    core = [verify_detatched_arm(), sample | Nock.Lib.logics_core()]
 
     sign = ECrypto.blood_l_signed_detached()
     valid = [sign, ECrypto.blood_msg() | ECrypto.londo().external.sign]
@@ -431,7 +431,7 @@ defmodule Examples.ENock do
   @spec bex() :: Noun.t()
   def bex() do
     sample = 888
-    core = [bex_arm(), sample | Nock.logics_core()]
+    core = [bex_arm(), sample | Nock.Lib.logics_core()]
 
     assert Nock.nock(core, [9, 2, 10, [6, 1 | 2], 0 | 1])
            |> elem(1)
@@ -467,7 +467,7 @@ defmodule Examples.ENock do
   @spec mix() :: Noun.t()
   def mix() do
     sample = [0 | 0]
-    core = [mix_arm(), sample | Nock.logics_core()]
+    core = [mix_arm(), sample | Nock.Lib.logics_core()]
 
     assert Nock.nock(core, [9, 2, 10, [6, 1, 3 | 5], 0 | 1])
            |> elem(1)
@@ -499,7 +499,7 @@ defmodule Examples.ENock do
   @spec mat() :: Noun.t()
   def mat() do
     sample = 0
-    core = [mat_arm(), sample | Nock.logics_core()]
+    core = [mat_arm(), sample | Nock.Lib.logics_core()]
 
     core
   end
@@ -522,7 +522,7 @@ defmodule Examples.ENock do
   @spec shax() :: Noun.t()
   def shax() do
     sample = 0
-    core = [shax_arm(), sample | Nock.logics_core()]
+    core = [shax_arm(), sample | Nock.Lib.logics_core()]
 
     assert Nock.nock(core, [9, 2, 0 | 1])
            |> elem(1)
@@ -559,7 +559,7 @@ defmodule Examples.ENock do
 
     sample = [0, 0]
 
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @doc """
@@ -596,7 +596,7 @@ defmodule Examples.ENock do
 
     sample = [0, 0]
 
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @doc """
@@ -634,7 +634,7 @@ defmodule Examples.ENock do
 
     sample = [0, 0]
 
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @doc """
@@ -673,7 +673,7 @@ defmodule Examples.ENock do
 
     sample = [0, 0]
 
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @doc """
@@ -708,7 +708,7 @@ defmodule Examples.ENock do
   def abs() do
     arm = abs_arm()
     sample = 888
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # abs(--0) == 0
     assert Nock.nock(core, [9, 2, 10, [6, 1 | 0], 0 | 1])
@@ -749,7 +749,7 @@ defmodule Examples.ENock do
   def dif() do
     arm = dif_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # --3 - -2 == --5
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [6 | 3]], 0 | 1])
@@ -785,7 +785,7 @@ defmodule Examples.ENock do
   def dul() do
     arm = dul_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # dul(-1, --5) == 9
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [1 | 10]], 0 | 1])
@@ -826,7 +826,7 @@ defmodule Examples.ENock do
   def fra() do
     arm = fra_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # -1 / -1 == --1
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [1 | 1]], 0 | 1])
@@ -872,7 +872,7 @@ defmodule Examples.ENock do
   def new() do
     arm = new_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # new(%.n, 2) == -2
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [1 | 2]], 0 | 1])
@@ -908,7 +908,7 @@ defmodule Examples.ENock do
   def old() do
     arm = old_arm()
     sample = 888
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # old(-2) == [%.n, 2]
     assert Nock.nock(core, [9, 2, 10, [6, 1 | 3], 0 | 1])
@@ -944,7 +944,7 @@ defmodule Examples.ENock do
   def pro() do
     arm = pro_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # -3 * --3 == -9
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [5 | 6]], 0 | 1])
@@ -980,7 +980,7 @@ defmodule Examples.ENock do
   def rem() do
     arm = rem_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # -17 % -3 == -2
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [33 | 5]], 0 | 1])
@@ -1026,7 +1026,7 @@ defmodule Examples.ENock do
   def sum() do
     arm = sum_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # -11 + --2 == -9
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [21 | 4]], 0 | 1])
@@ -1050,7 +1050,7 @@ defmodule Examples.ENock do
   def sun() do
     arm = sun_arm()
     sample = 888
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # sun(90) == 180
     assert Nock.nock(core, [9, 2, 10, [6, 1 | 90], 0 | 1])
@@ -1069,7 +1069,7 @@ defmodule Examples.ENock do
   def syn() do
     arm = syn_arm()
     sample = 888
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # syn(--0) == %.y
     assert Nock.nock(core, [9, 2, 10, [6, 1 | 0], 0 | 1])
@@ -1110,7 +1110,7 @@ defmodule Examples.ENock do
   def cmp() do
     arm = cmp_arm()
     sample = [888 | 999]
-    core = [arm, sample | Nock.logics_core()]
+    core = [arm, sample | Nock.Lib.logics_core()]
 
     # cmp(-2, --1) == -1
     assert Nock.nock(core, [9, 2, 10, [6, 1 | [3 | 2]], 0 | 1])
@@ -1155,7 +1155,7 @@ defmodule Examples.ENock do
   @spec lte() :: Noun.t()
   def lte() do
     sample = [888 | 999]
-    core = [lte_arm(), sample | Nock.logics_core()]
+    core = [lte_arm(), sample | Nock.Lib.logics_core()]
 
     max_test_val = 4
 
@@ -1177,7 +1177,7 @@ defmodule Examples.ENock do
 
   def delta_add_call(delta1, delta2) do
     sample = [delta1 | delta2]
-    [delta_add_arm(), sample | Nock.logics_core()]
+    [delta_add_arm(), sample | Nock.Lib.logics_core()]
   end
 
   def delta_add_test() do
@@ -1196,7 +1196,7 @@ defmodule Examples.ENock do
 
   def delta_sub_call(delta1, delta2) do
     sample = [delta1 | delta2]
-    [delta_sub_arm(), sample | Nock.logics_core()]
+    [delta_sub_arm(), sample | Nock.Lib.logics_core()]
   end
 
   def delta_sub_test() do
@@ -1215,7 +1215,7 @@ defmodule Examples.ENock do
 
   def action_delta_call(action) do
     sample = action
-    [action_delta_arm(), sample | Nock.logics_core()]
+    [action_delta_arm(), sample | Nock.Lib.logics_core()]
   end
 
   def action_delta_test() do
@@ -1234,7 +1234,7 @@ defmodule Examples.ENock do
 
   def make_delta_call(actions) do
     sample = actions
-    [make_delta_arm(), sample | Nock.logics_core()]
+    [make_delta_arm(), sample | Nock.Lib.logics_core()]
   end
 
   def make_delta_test() do
@@ -1255,7 +1255,7 @@ defmodule Examples.ENock do
 
   def make_is_commitment_call(atom) do
     sample = atom
-    [is_commitment_arm(), sample | Nock.logics_core()]
+    [is_commitment_arm(), sample | Nock.Lib.logics_core()]
   end
 
   def is_commitment_test() do
@@ -1286,7 +1286,7 @@ defmodule Examples.ENock do
 
   def make_is_nullifier_call(atom) do
     sample = atom
-    [is_nullifier_arm(), sample | Nock.logics_core()]
+    [is_nullifier_arm(), sample | Nock.Lib.logics_core()]
   end
 
   def is_nullifier_test() do
@@ -1594,7 +1594,7 @@ defmodule Examples.ENock do
   def og_arm() do
     arm = "[8 [9 47 0 63] 10 [6 0 14] 0 2]" |> Noun.Format.parse_always()
     sample = 0
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @spec og_call(non_neg_integer()) :: :error | {:ok, Noun.t()}
@@ -1620,7 +1620,7 @@ defmodule Examples.ENock do
 
     sample = [0, 0]
 
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @spec raws_with_core_call(non_neg_integer(), non_neg_integer()) ::
@@ -1650,7 +1650,7 @@ defmodule Examples.ENock do
   def split_arm() do
     arm = "[7 [0 6] 9 21 0 1]" |> Noun.Format.parse_always()
     sample = 0
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @spec split_call(Noun.t()) :: :error | {:ok, Noun.t()}
@@ -1712,7 +1712,7 @@ defmodule Examples.ENock do
   @spec factorial() :: Noun.t()
   def factorial() do
     sample = 1
-    core = [factorial_arm(), sample | Nock.logics_core()]
+    core = [factorial_arm(), sample | Nock.Lib.logics_core()]
 
     assert Nock.nock(core, [9, 2, 10, [6, 1 | 7], 0 | 1])
            |> elem(1)
@@ -1745,7 +1745,7 @@ defmodule Examples.ENock do
       )
 
     sample = [999 | 888]
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @spec block_calling_mono(Noun.t(), Noun.t()) :: Noun.t()
@@ -1767,14 +1767,14 @@ defmodule Examples.ENock do
       )
 
     sample = 999
-    [arm, sample | Nock.logics_core()]
+    [arm, sample | Nock.Lib.logics_core()]
   end
 
   @spec increment_counter_val(Noun.t()) :: Noun.t()
   def increment_counter_val(val) do
     arm = [[1 | val], 4, 12, [1 | 0], [0 | 6], 1, val | 0]
     sample = 0
-    [[8, [1 | sample], [1 | arm], 0 | 1] | Nock.logics_core()]
+    [[8, [1 | sample], [1 | arm], 0 | 1] | Nock.Lib.logics_core()]
   end
 
   # [%ctr 0]
@@ -1782,7 +1782,7 @@ defmodule Examples.ENock do
   def zero_counter(val) do
     arm = [1, val | 0]
     sample = 0
-    [[8, [1 | sample], [1 | arm], 0 | 1] | Nock.logics_core()]
+    [[8, [1 | sample], [1 | arm], 0 | 1] | Nock.Lib.logics_core()]
   end
 
   ####################################################################
