+++ apps/anoma_lib/lib/nock/lib.ex
@@ -1,0 +1,2489 @@
+defmodule Nock.Lib do
+  @layers 9
+
+  @doc """
+  Gives the total numbers of layers in the standard library
+  """
+  @spec stdlib_layers() :: non_neg_integer()
+  def stdlib_layers, do: @layers
+
+  # compiled and evaluated output from hoon/anoma.hoon.
+  # eventually, will be replaced with merely compiled output, with
+  # evaluation to take place during nock vm bootstrapping.
+  stdlib_string = """
+  [ [ [ 7
+      [ 8
+        [1 0 0]
+        [ 1
+          8
+          [[8 [9 756 0 7] 9 2 10 [6 0 28] 0 2] 8 [9 756 0 7] 9 2 10 [6 0 29] 0 2]
+          6
+          [0 8]
+          [ 6
+            [0 10]
+            [8 [9 758 0 15] 9 2 10 [6 [7 [0 3] 1 0] 7 [0 3] 8 [9 20 0 2.047] 9 2 10 [6 [0 25] 0 27] 0 2] 0 2]
+            6
+            [8 [9 22 0 2.047] 9 2 10 [6 [0 25] 0 27] 0 2]
+            [8 [9 758 0 15] 9 2 10 [6 [7 [0 3] 1 0] 7 [0 3] 8 [9 47 0 2.047] 9 2 10 [6 [0 25] 0 27] 0 2] 0 2]
+            8
+            [9 758 0 15]
+            9
+            2
+            10
+            [6 [7 [0 3] 1 1] 7 [0 3] 8 [9 47 0 2.047] 9 2 10 [6 [0 27] 0 25] 0 2]
+            0
+            2
+          ]
+          6
+          [0 10]
+          [ 6
+            [8 [9 22 0 2.047] 9 2 10 [6 [0 25] 0 27] 0 2]
+            [8 [9 758 0 15] 9 2 10 [6 [7 [0 3] 1 1] 7 [0 3] 8 [9 47 0 2.047] 9 2 10 [6 [0 25] 0 27] 0 2] 0 2]
+            8
+            [9 758 0 15]
+            9
+            2
+            10
+            [6 [7 [0 3] 1 0] 7 [0 3] 8 [9 47 0 2.047] 9 2 10 [6 [0 27] 0 25] 0 2]
+            0
+            2
+          ]
+          8
+          [9 758 0 15]
+          9
+          2
+          10
+          [6 [7 [0 3] 1 1] 7 [0 3] 8 [9 20 0 2.047] 9 2 10 [6 [0 25] 0 27] 0 2]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.173.491 [0 7] 0]
+      0
+      1
+    ]
+    [7 [8 [1 0] [1 8 [9 4 0 1.023] 9 2 10 [6 [7 [0 3] 1 2] 0 14] 0 2] 0 1] 11 [1.953.718.630 1 7.239.027 [0 7] 0] 0 1]
+    [ 7
+      [ 8
+        [1 0 0]
+        [ 1
+          8
+          [8 [9 756 0 7] 9 2 10 [6 0 28] 0 2]
+          6
+          [0 4]
+          [8 [9 46 0 2.047] 9 2 10 [6 [0 13] 0 61] 0 2]
+          8
+          [9 47 0 2.047]
+          9
+          2
+          10
+          [6 [0 61] 0 13]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.107.940 [0 7] 0]
+      0
+      1
+    ]
+    [ 7
+      [ 8
+        [1 0 0]
+        [ 1
+          8
+          [9 758 0 7]
+          9
+          2
+          10
+          [ 6
+            [7 [0 3] 5 [1 0] 8 [9 4 0 63] 9 2 10 [6 [7 [0 3] 8 [9 188 0 7] 9 2 10 [6 0 28] 0 2] 7 [0 3] 8 [9 188 0 7] 9 2 10 [6 0 29] 0 2] 0 2]
+            7
+            [0 3]
+            8
+            [9 4 0 1.023]
+            9
+            2
+            10
+            [6 [7 [0 3] 8 [9 1.515 0 7] 9 2 10 [6 0 28] 0 2] 7 [0 3] 8 [9 1.515 0 7] 9 2 10 [6 0 29] 0 2]
+            0
+            2
+          ]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.303.792 [0 7] 0]
+      0
+      1
+    ]
+    [ [ 7
+        [8 [1 0] [1 5 [1 0] 8 [7 [7 [0 7] 9 4 0 3] 9 1.375 0 1] 9 2 10 [6 [7 [0 3] 1 0] 0 14] 0 2] 0 1]
+        11
+        [1.953.718.630 1 7.240.051 [0 7] 0]
+        0
+        1
+      ]
+      [ [7 [8 [1 0] [1 [8 [9 188 0 7] 9 2 10 [6 0 14] 0 2] 8 [9 1.515 0 7] 9 2 10 [6 0 14] 0 2] 0 1] 11 [1.953.718.630 1 6.581.359 [0 7] 0] 0 1]
+        [ 7
+          [ 8
+            [1 0 0]
+            [1 8 [9 759 0 7] 9 2 10 [6 [0 28] 7 [0 3] 8 [9 46 0 7] 9 2 10 [6 [0 29] 7 [0 3] 8 [9 190 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2] 0 2]
+            0
+            1
+          ]
+          11
+          [1.953.718.630 1 7.169.394 [0 7] 0]
+          0
+          1
+        ]
+        7
+        [ 8
+          [1 0]
+          [ 1
+            8
+            [7 [7 [0 7] 9 4 0 3] 9 20 0 255]
+            9
+            2
+            10
+            [ 6
+              [7 [0 3] 8 [7 [7 [0 7] 9 4 0 3] 9 1.375 0 1] 9 2 10 [6 [7 [0 3] 1 0] 0 14] 0 2]
+              7
+              [0 3]
+              8
+              [7 [7 [0 7] 9 4 0 3] 9 87 0 1]
+              9
+              2
+              10
+              [6 [7 [0 3] 1 0] 0 14]
+              0
+              2
+            ]
+            0
+            2
+          ]
+          0
+          1
+        ]
+        11
+        [1.953.718.630 1 7.561.825 [0 7] 0]
+        0
+        1
+      ]
+      [ 7
+        [ 8
+          [1 0 0]
+          [ 1
+            6
+            [0 12]
+            [8 [9 4 0 1.023] 9 2 10 [6 [7 [0 3] 1 2] 0 29] 0 2]
+            6
+            [5 [1 0] 0 13]
+            [1 0]
+            4
+            8
+            [9 4 0 1.023]
+            9
+            2
+            10
+            [6 [7 [0 3] 1 2] 7 [0 3] 8 [9 342 0 1.023] 9 2 10 [6 0 29] 0 2]
+            0
+            2
+          ]
+          0
+          1
+        ]
+        11
+        [1.953.718.630 1 7.824.750 [0 7] 0]
+        0
+        1
+      ]
+      7
+      [ 8
+        [1 0 0]
+        [ 1
+          8
+          [9 4 0 7]
+          9
+          2
+          10
+          [ 6
+            [0 28]
+            7
+            [0 3]
+            8
+            [9 758 0 7]
+            9
+            2
+            10
+            [6 [7 [0 3] 6 [8 [9 188 0 7] 9 2 10 [6 0 29] 0 2] [1 1] 1 0] 7 [0 3] 8 [9 1.515 0 7] 9 2 10 [6 0 29] 0 2]
+            0
+            2
+          ]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 6.711.652 [0 7] 0]
+      0
+      1
+    ]
+    [ 7
+      [ 8
+        [1 0 0]
+        [ 1
+          8
+          [9 758 0 7]
+          9
+          2
+          10
+          [ 6
+            [7 [0 3] 5 [1 0] 8 [9 4 0 63] 9 2 10 [6 [7 [0 3] 8 [9 188 0 7] 9 2 10 [6 0 28] 0 2] 7 [0 3] 8 [9 188 0 7] 9 2 10 [6 0 29] 0 2] 0 2]
+            7
+            [0 3]
+            8
+            [9 170 0 1.023]
+            9
+            2
+            10
+            [6 [7 [0 3] 8 [9 1.515 0 7] 9 2 10 [6 0 28] 0 2] 7 [0 3] 8 [9 1.515 0 7] 9 2 10 [6 0 29] 0 2]
+            0
+            2
+          ]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 6.386.278 [0 7] 0]
+      0
+      1
+    ]
+    7
+    [ 8
+      [1 0 0]
+      [ 1
+        6
+        [5 [0 12] 0 13]
+        [1 0]
+        6
+        [8 [9 188 0 7] 9 2 10 [6 0 28] 0 2]
+        [6 [8 [9 188 0 7] 9 2 10 [6 0 29] 0 2] [6 [8 [9 43 0 1.023] 9 2 10 [6 [0 28] 0 29] 0 2] [1 2] 1 1] 1 2]
+        6
+        [8 [9 188 0 7] 9 2 10 [6 0 29] 0 2]
+        [1 1]
+        6
+        [8 [9 43 0 1.023] 9 2 10 [6 [0 28] 0 29] 0 2]
+        [1 1]
+        1
+        2
+      ]
+      0
+      1
+    ]
+    11
+    [1.953.718.630 1 7.368.035 [0 7] 0]
+    0
+    1
+  ]
+  [ [ [ 1
+        [ 7
+          [ 8
+            [1 0 0]
+            [ 1
+              8
+              [1 0 0]
+              8
+              [ 1
+                6
+                [6 [5 [1 0] 0 60] [1 0] 6 [5 [1 0] 0 61] [1 0] 1 1]
+                [0 13]
+                9
+                2
+                10
+                [30 [8 [9 87 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 124] 0 2] 8 [9 87 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 125] 0 2]
+                10
+                [ 6
+                  [4 0 12]
+                  8
+                  [9 20 0 4.095]
+                  9
+                  2
+                  10
+                  [ 6
+                    [0 29]
+                    7
+                    [0 3]
+                    8
+                    [9 341 0 31]
+                    9
+                    2
+                    10
+                    [ 6
+                      [7 [0 3] [1 0] 0 12]
+                      7
+                      [0 3]
+                      6
+                      [5 [1 0] 8 [9 1.375 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 124] 0 2]
+                      [1 0]
+                      6
+                      [5 [1 0] 8 [9 1.375 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 125] 0 2]
+                      [1 0]
+                      1
+                      1
+                    ]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                0
+                1
+              ]
+              9
+              2
+              0
+              1
+            ]
+            0
+            1
+          ]
+          11
+          [1.953.718.630 1 7.563.620 [0 7] 0]
+          0
+          1
+        ]
+        [ [ 7
+            [ 8
+              [1 0 0]
+              [ 1
+                8
+                [1 0 0]
+                8
+                [ 1
+                  6
+                  [6 [5 [1 0] 0 60] [6 [5 [1 0] 0 61] [1 0] 1 1] 1 1]
+                  [0 13]
+                  9
+                  2
+                  10
+                  [30 [8 [9 87 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 124] 0 2] 8 [9 87 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 125] 0 2]
+                  10
+                  [ 6
+                    [4 0 12]
+                    8
+                    [9 20 0 4.095]
+                    9
+                    2
+                    10
+                    [ 6
+                      [0 29]
+                      7
+                      [0 3]
+                      8
+                      [9 341 0 31]
+                      9
+                      2
+                      10
+                      [ 6
+                        [7 [0 3] [1 0] 0 12]
+                        7
+                        [0 3]
+                        6
+                        [5 [1 0] 8 [9 1.375 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 124] 0 2]
+                        [6 [5 [1 0] 8 [9 1.375 0 31] 9 2 10 [6 [7 [0 3] 1 0] 0 125] 0 2] [1 0] 1 1]
+                        1
+                        1
+                      ]
+                      0
+                      2
+                    ]
+                    0
+                    2
+                  ]
+                  0
+                  1
+                ]
+                9
+                2
+                0
+                1
+              ]
+              0
+              1
+            ]
+            11
+            [1.953.718.630 1 7.237.475 [0 7] 0]
+            0
+            1
+          ]
+          [ [ 8
+              [1 0]
+              [ 1
+                8
+                [ 6
+                  [3 0 6]
+                  [[8 [7 [0 7] 9 342 0 1] 9 2 10 [6 0 28] 0 2] 8 [7 [0 7] 9 1.374 0 1] 9 2 10 [6 0 29] 0 2]
+                  8
+                  [7 [0 7] 9 342 0 1]
+                  9
+                  2
+                  10
+                  [6 0 14]
+                  0
+                  2
+                ]
+                8
+                [5 [0 14] 0 2]
+                0
+                6
+              ]
+              0
+              1
+            ]
+            [ [ 7
+                [ 8
+                  [1 0 0]
+                  [1 6 [5 [1 0] 0 13] [1 0] [8 [9 1.375 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] 9 2 10 [13 8 [9 87 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] 0 1]
+                  0
+                  1
+                ]
+                11
+                [1.953.718.630 1 7.367.026 [0 7] 0]
+                0
+                1
+              ]
+              7
+              [ 8
+                [1 0 0]
+                [1 8 [6 [6 [3 0 12] [1 1] 1 0] [[0 12] 1 1] 0 12] 8 [8 [9 10 0 255] 9 90 10 [6 0 12] 0 2] 9 2 10 [6 [0 13] 0 61] 0 2]
+                0
+                1
+              ]
+              11
+              [1.953.718.630 1 6.845.292 [0 7] 0]
+              0
+              1
+            ]
+            [8 [1 0] [1 8 [6 [6 [3 0 6] [1 1] 1 0] [0 6] 0 0] 8 [5 [0 14] 0 2] 0 6] 0 1]
+            [8 [1 0 0] [1 8 [8 [9 10 0 127] 9 190 10 [6 0 28] 0 2] 9 2 10 [6 0 29] 0 2] 0 1]
+            [8 [1 0] [1 8 [7 [0 7] 1 1] 8 [5 [0 14] 0 2] 0 6] 0 1]
+            7
+            [ 8
+              [1 0 0]
+              [1 8 [6 [6 [3 0 12] [1 1] 1 0] [[0 12] 1 1] 0 12] 8 [8 [9 10 0 255] 9 367 10 [6 0 12] 0 2] 9 2 10 [6 [0 13] 0 61] 0 2]
+              0
+              1
+            ]
+            11
+            [1.953.718.630 1 6.581.861 [0 7] 0]
+            0
+            1
+          ]
+          [ 7
+            [ 8
+              [1 0 [1 1] 0]
+              [ 1
+                8
+                [8 [9 10 0 127] 9 367 10 [6 0 28] 0 2]
+                9
+                2
+                10
+                [6 [0 117] 7 [0 3] 8 [8 [9 10 0 127] 9 767 10 [6 0 28] 0 2] 9 2 10 [6 [0 116] 0 59] 0 2]
+                0
+                2
+              ]
+              0
+              1
+            ]
+            11
+            [1.953.718.630 1 7.632.227 [0 7] 0]
+            0
+            1
+          ]
+          7
+          [8 [1 0 0] [1 8 [6 [6 [3 0 12] [1 1] 1 0] [[0 12] 1 1] 0 12] 8 [8 [9 10 0 255] 9 767 10 [6 0 12] 0 2] 9 2 10 [6 [0 13] 0 61] 0 2] 0 1]
+          11
+          [1.953.718.630 1 6.845.298 [0 7] 0]
+          0
+          1
+        ]
+        [ 8
+          [1 0]
+          [ 1
+            [8 [1 0 0] [1 8 [9 44 0 7] 9 2 10 [6 7 [0 3] 8 [9 20 0 4.095] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2] 0 1]
+            [8 [9 4 0 127] 9 2 10 [6 7 [0 3] 8 [9 4 0 127] 9 2 10 [6 0 14] 0 2] 0 2]
+            [ [8 [1 0] [1 8 [9 1.375 0 31] 9 2 10 [6 [0 62] 0 14] 0 2] 0 1]
+              8
+              [1 0]
+              [ 1
+                8
+                [9 47 0 4.095]
+                9
+                2
+                10
+                [6 [7 [0 3] 8 [9 342 0 4.095] 9 2 10 [6 7 [0 3] 9 10 0 7] 0 2] 7 [0 3] 8 [9 44 0 7] 9 2 10 [6 0 14] 0 2]
+                0
+                2
+              ]
+              0
+              1
+            ]
+            [ 8
+              [1 0 0 0]
+              [ 1
+                8
+                [8 [9 44 0 7] 9 2 10 [6 0 59] 0 2]
+                8
+                [8 [9 4 0 1.023] 9 2 10 [6 7 [0 3] 8 [9 47 0 8.191] 9 2 10 [6 [0 126] 0 60] 0 2] 0 2]
+                8
+                [8 [9 46 0 16.383] 9 2 10 [6 [0 250] 0 6] 0 2]
+                8
+                [9 44 0 63]
+                9
+                2
+                10
+                [ 6
+                  7
+                  [0 3]
+                  8
+                  [9 20 0 255]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 8 [9 87 0 255] 9 2 10 [6 [7 [0 3] [0 124] 0 2] 0 30] 0 2]
+                    7
+                    [0 3]
+                    8
+                    [9 341 0 255]
+                    9
+                    2
+                    10
+                    [6 [7 [0 3] [0 124] 8 [9 47 0 32.767] 9 2 10 [6 [0 14] 0 6] 0 2] 0 30]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              0
+              1
+            ]
+            8
+            [1 0]
+            [ 1
+              7
+              [10 [6 8 [9 44 0 7] 9 2 10 [6 0 14] 0 2] 0 1]
+              6
+              [8 [9 84 0 4.095] 9 2 10 [6 [0 62] 7 [0 3] 1 3] 0 2]
+              [0 6]
+              8
+              [8 [9 342 0 4.095] 9 2 10 [6 0 62] 0 2]
+              8
+              [9 20 0 63]
+              9
+              2
+              10
+              [ 6
+                [ 7
+                  [0 3]
+                  8
+                  [9 341 0 63]
+                  9
+                  2
+                  10
+                  [6 [0 6] 7 [0 3] 9 2 10 [30 0 2] 10 [6 8 [9 86 0 63] 9 2 10 [6 [0 6] [7 [0 3] 1 0 1] 0 30] 0 2] 0 3]
+                  0
+                  2
+                ]
+                7
+                [0 3]
+                9
+                2
+                10
+                [30 0 2]
+                10
+                [6 8 [9 86 0 63] 9 2 10 [6 [0 6] [7 [0 3] 1 1 1] 0 30] 0 2]
+                0
+                3
+              ]
+              0
+              2
+            ]
+            0
+            1
+          ]
+          0
+          1
+        ]
+        [ 7
+          [ 8
+            [1 0 0]
+            [ 1
+              6
+              [5 [1 0] 0 13]
+              [1 0]
+              8
+              [9 20 0 1.023]
+              9
+              2
+              10
+              [ 6
+                [7 [0 3] 8 [9 1.375 0 7] 9 2 10 [6 [7 [0 3] [0 12] 0 52] 0 117] 0 2]
+                7
+                [0 3]
+                8
+                [9 341 0 7]
+                9
+                2
+                10
+                [6 [7 [0 3] [0 12] 0 52] 7 [0 3] 9 2 10 [13 0 27] 0 1]
+                0
+                2
+              ]
+              0
+              2
+            ]
+            0
+            1
+          ]
+          11
+          [1.953.718.630 1 7.233.891 [0 7] 0]
+          0
+          1
+        ]
+        [ 7
+          [ 8
+            [1 0 0 [0 15] 0 0 0]
+            [1 8 [9 95 0 7] 9 2 10 [6 [0 28] 7 [0 3] 8 [9 22 0 511] 9 2 10 [6 [7 [0 3] 8 [9 340 0 7] 9 2 10 [6 [0 28] 0 58] 0 2] 0 59] 0 2] 0 2]
+            0
+            1
+          ]
+          11
+          [1.953.718.630 1 7.239.026 [0 7] 0]
+          0
+          1
+        ]
+        7
+        [ 8
+          [1 0 0]
+          [ 1
+            8
+            [6 [6 [3 0 12] [1 1] 1 0] [[0 12] 1 1] 0 12]
+            8
+            [1 0]
+            8
+            [ 1
+              6
+              [5 [1 0] 0 125]
+              [1 0]
+              8
+              [9 20 0 8.191]
+              9
+              2
+              10
+              [ 6
+                [7 [0 3] 9 2 10 [125 0 251] 10 [6 4 0 6] 0 1]
+                7
+                [0 3]
+                8
+                [9 341 0 63]
+                9
+                2
+                10
+                [ 6
+                  [7 [0 3] [0 28] 8 [9 4 0 8.191] 9 2 10 [6 [0 61] 0 14] 0 2]
+                  7
+                  [0 3]
+                  8
+                  [9 1.375 0 63]
+                  9
+                  2
+                  10
+                  [6 [7 [0 3] [0 28] 0 29] 0 506]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              0
+              2
+            ]
+            9
+            2
+            0
+            1
+          ]
+          0
+          1
+        ]
+        11
+        [1.953.718.630 1 7.366.002 [0 7] 0]
+        0
+        1
+      ]
+      0
+      1
+    ]
+    [ 7
+      [8 [1 0 0] [1 8 [9 22 0 7] 9 2 10 [6 7 [0 3] 8 [9 4 0 31] 9 2 10 [6 [0 28] 7 [0 3] 8 [9 22 0 7] 9 2 10 [6 0 29] 0 2] 0 2] 0 2] 0 1]
+      11
+      [1.953.718.630 1 1.935.763.571 [0 7] 0]
+      0
+      1
+    ]
+    [ 7
+      [ 8
+        [1 0]
+        [1 8 [7 [7 [0 1] 9 4 0 7] 9 46 0 3] 9 2 10 [6 7 [0 3] [8 [7 [7 [0 1] 9 4 0 7] 9 686 0 1] 9 2 10 [6 [7 [0 3] 1 3] 0 14] 0 2] 0 6] 0 2]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 2.019.649.651 [0 7] 0]
+      0
+      1
+    ]
+    [ 7
+      [ 8
+        [1 0 0]
+        [ 1
+          7
+          [10 [13 8 [7 [7 [0 1] 9 4 0 7] 9 86 0 1] 9 2 10 [6 [7 [0 3] 1 3] [7 [0 3] [1 0] 0 12] 0 29] 0 2] 0 1]
+          8
+          [ [7 [7 [7 [0 1] 9 4 0 7] 9 22 0 1] 10 [6 1 5] 0 1]
+            8
+            [1 0 0]
+            [1 8 [7 [7 [0 7] 9 4 0 7] 9 86 0 1] 9 2 10 [6 [7 [0 3] 1 5] [7 [0 3] [0 12] 1 1] 0 29] 0 2]
+            0
+            1
+          ]
+          8
+          [[9 4 0 4] [9 46 0 4] [9 47 0 4] 9 45 0 4]
+          8
+          [8 [7 [7 [0 7] 9 4 0 7] 9 341 0 1] 9 2 10 [6 [7 [0 3] 1 0 3] 0 124] 0 2]
+          8
+          [ 8
+            [7 [7 [0 15] 9 4 0 7] 9 46 0 1]
+            9
+            2
+            10
+            [ 6
+              [7 [0 3] 1 0]
+              7
+              [0 3]
+              [[0 2] 0 125]
+              [1 8 128]
+              [ [ 8
+                  [7 [7 [0 15] 9 4 0 7] 9 46 0 255]
+                  9
+                  2
+                  10
+                  [ 6
+                    [ 7
+                      [0 3]
+                      8
+                      [7 [7 [0 15] 9 4 0 7] 9 47 0 255]
+                      9
+                      2
+                      10
+                      [ 6
+                        [7 [0 3] 1 960]
+                        7
+                        [0 3]
+                        8
+                        [7 [7 [0 15] 9 4 0 7] 9 46 0 255]
+                        9
+                        2
+                        10
+                        [6 [7 [0 3] 8 [7 [7 [0 15] 9 4 0 7] 9 20 0 255] 9 2 10 [6 [7 [0 3] 1 8] 0 6] 0 2] 7 [0 3] 1 512]
+                        0
+                        2
+                      ]
+                      0
+                      2
+                    ]
+                    7
+                    [0 3]
+                    1
+                    512
+                  ]
+                  0
+                  2
+                ]
+                1
+                0
+              ]
+              [[1 64] 8 [8 [7 [7 [0 15] 9 4 0 7] 9 22 0 1] 9 47 10 [6 7 [0 3] 1 6] 0 2] 9 2 10 [6 0 6] 0 2]
+              1
+              0
+            ]
+            0
+            2
+          ]
+          8
+          [8 [7 [7 [0 31] 9 4 0 7] 9 686 0 1] 9 2 10 [6 [7 [0 3] 1 9] 0 6] 0 2]
+          8
+          [ 1
+            25.051.139.735.836.467.913.601.071.899.189.108.501.681.633.092.613.316.132.753.366.958.947.502.841.281.110.980.347.811.086.624.969.305.314.199.562.795.868.290.539.880.502.533.646.505.489.797.110.349.007.385.020.507.326.982.043.050.618.112.420.254.613.810.441.709.594.605.123.090.411.975.756.494.285.771.699.200.369.901.479.195.251.226.368.983.824.020.564.277.645.963.860.728.452.821.576.845.901.498.668.417.244.438.721.537.663.670.541.944.820.957.180.957.595.559.282.976.806.173.113.161.068.298.822.071.065.329.290.006.052.849.814.285.001.949.914.564.097.058.408.480.133.985.233.335.799.884.203.712.730.341.384.999.677.089.997.083.749.077.591.931.498.939.520.449.886.954.646.413.138.343.858.395.935.213.418.018.409.268.340.744.776.361.518.554.939.863.400.075.967.197.509.182.087.778.881.547.827.184.266.701.615.699.472.280
+          ]
+          8
+          [1 41.557.658.498.906.279.274.860.226.408.925.318.911.382.702.236.748.615.442.085.426.012.007.055.353.447]
+          8
+          [1 0]
+          8
+          [ 1
+            6
+            [5 [0 6] 0 62]
+            [8 [7 [7 [0 1.023] 9 4 0 7] 9 94 0 1] 9 2 10 [6 [7 [0 3] 1 5] [0 30] 0 8.182] 0 2]
+            8
+            [ 8
+              [8 [7 [7 [0 1.023] 9 4 0 7] 9 86 0 1] 9 2 10 [6 [7 [0 3] 1 9] [7 [0 3] [0 6] 1 1] 0 254] 0 2]
+              8
+              [8 [7 [7 [0 2.047] 9 4 0 7] 9 94 0 1] 9 2 10 [6 [7 [0 3] 1 5] [0 6] 0 16.374] 0 2]
+              8
+              [1 16]
+              8
+              [ 1
+                6
+                [5 [1 64] 0 6]
+                [0 14]
+                8
+                [ [8 [0 32.765] 9 2 10 [6 [7 [0 3] 8 [7 [7 [0 16.383] 9 4 0 7] 9 47 0 255] 9 2 10 [6 [0 14] 7 [0 3] 1 15] 0 2] 0 30] 0 2]
+                  [8 [0 32.765] 9 2 10 [6 [7 [0 3] 8 [7 [7 [0 16.383] 9 4 0 7] 9 47 0 255] 9 2 10 [6 [0 14] 7 [0 3] 1 2] 0 2] 0 30] 0 2]
+                  [8 [0 32.765] 9 2 10 [6 [7 [0 3] 8 [7 [7 [0 16.383] 9 4 0 7] 9 47 0 255] 9 2 10 [6 [0 14] 7 [0 3] 1 16] 0 2] 0 30] 0 2]
+                  8
+                  [0 32.765]
+                  9
+                  2
+                  10
+                  [6 [7 [0 3] 8 [7 [7 [0 16.383] 9 4 0 7] 9 47 0 255] 9 2 10 [6 [0 14] 7 [0 3] 1 7] 0 2] 0 30]
+                  0
+                  2
+                ]
+                8
+                [ 8
+                  [7 [7 [0 32.767] 9 4 0 7] 9 4 0 15]
+                  8
+                  [0 2]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 7 [0 3] 8 [0 65.530] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 7] 0 12] 0 2]
+                    7
+                    [0 3]
+                    8
+                    [0 2]
+                    9
+                    2
+                    10
+                    [ 6
+                      [7 [0 3] 7 [0 3] 8 [0 65.530] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 18] 0 12] 0 2]
+                      7
+                      [0 3]
+                      7
+                      [0 3]
+                      8
+                      [7 [7 [0 32.767] 9 4 0 7] 9 87 0 1]
+                      9
+                      2
+                      10
+                      [6 [7 [0 3] 1 0 3] 0 12]
+                      0
+                      2
+                    ]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                8
+                [ 8
+                  [7 [7 [0 65.535] 9 4 0 7] 9 4 0 15]
+                  8
+                  [0 2]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 7 [0 3] 8 [0 131.066] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 17] 0 58] 0 2]
+                    7
+                    [0 3]
+                    8
+                    [0 2]
+                    9
+                    2
+                    10
+                    [ 6
+                      [7 [0 3] 7 [0 3] 8 [0 131.066] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 19] 0 58] 0 2]
+                      7
+                      [0 3]
+                      7
+                      [0 3]
+                      8
+                      [7 [7 [0 65.535] 9 4 0 7] 9 87 0 1]
+                      9
+                      2
+                      10
+                      [6 [7 [0 3] 1 0 10] 0 58]
+                      0
+                      2
+                    ]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                8
+                [8 [0 131.068] 8 [0 2] 9 2 10 [6 [0 502] 7 [0 3] 8 [0 2] 9 2 10 [6 [0 30] 7 [0 3] 8 [0 2] 9 2 10 [6 [0 503] 0 14] 0 2] 0 2] 0 2]
+                9
+                2
+                10
+                [ 14
+                  8
+                  [7 [7 [0 262.143] 9 4 0 7] 9 20 0 1]
+                  9
+                  2
+                  10
+                  [6 [7 [0 3] 8 [7 [7 [0 262.143] 9 4 0 7] 9 341 0 1] 9 2 10 [6 [7 [0 3] [1 5] 0 126] 0 6] 0 2] 0 510]
+                  0
+                  2
+                ]
+                10
+                [6 4 0 126]
+                0
+                31
+              ]
+              9
+              2
+              0
+              1
+            ]
+            8
+            [1 0]
+            8
+            [ [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 0] 0 126] 0 2]
+              [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 1] 0 126] 0 2]
+              [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 2] 0 126] 0 2]
+              [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 3] 0 126] 0 2]
+              [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 4] 0 126] 0 2]
+              [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 5] 0 126] 0 2]
+              [8 [0 8.189] 9 2 10 [6 [7 [0 3] 1 6] 0 126] 0 2]
+              8
+              [0 8.189]
+              9
+              2
+              10
+              [6 [7 [0 3] 1 7] 0 126]
+              0
+              2
+            ]
+            8
+            [ 1
+              6
+              [5 [1 64] 0 14]
+              [ 9
+                2
+                10
+                [ 14
+                  8
+                  [7 [7 [0 16.383] 9 4 0 7] 9 95 0 1]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 1 5]
+                    7
+                    [0 3]
+                    [8 [0 16.380] 9 2 10 [6 [0 28] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 0] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 58] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 1] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 118] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 2] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 238] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 3] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 478] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 4] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 958] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 5] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 1.918] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 6] 0 510] 0 2] 0 2]
+                    [8 [0 16.380] 9 2 10 [6 [0 1.919] 7 [0 3] 8 [0 32.765] 9 2 10 [6 [7 [0 3] 1 7] 0 510] 0 2] 0 2]
+                    1
+                    0
+                  ]
+                  0
+                  2
+                ]
+                10
+                [6 4 0 126]
+                0
+                31
+              ]
+              8
+              [ 8
+                [7 [7 [0 16.383] 9 4 0 7] 9 4 0 15]
+                8
+                [0 2]
+                9
+                2
+                10
+                [ 6
+                  [7 [0 3] 7 [0 3] 8 [0 32.762] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 2] 0 28] 0 2]
+                  7
+                  [0 3]
+                  8
+                  [0 2]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 7 [0 3] 8 [0 32.762] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 13] 0 28] 0 2]
+                    7
+                    [0 3]
+                    7
+                    [0 3]
+                    8
+                    [0 32.762]
+                    9
+                    2
+                    10
+                    [6 [7 [0 3] 1 0] [7 [0 3] 1 22] 0 28]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              8
+              [ 8
+                [7 [7 [0 32.767] 9 4 0 7] 9 4 0 15]
+                8
+                [0 2]
+                9
+                2
+                10
+                [ 6
+                  [7 [0 3] 7 [0 3] 8 [7 [7 [0 32.767] 9 4 0 7] 9 4 0 1] 9 2 10 [6 [0 60] 0 122] 0 2]
+                  7
+                  [0 3]
+                  8
+                  [0 2]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 7 [0 3] 8 [7 [7 [0 32.767] 9 4 0 7] 9 4 0 1] 9 2 10 [6 [0 60] 0 246] 0 2]
+                    7
+                    [0 3]
+                    7
+                    [0 3]
+                    8
+                    [7 [7 [0 32.767] 9 4 0 7] 9 4 0 1]
+                    9
+                    2
+                    10
+                    [6 [0 122] 0 246]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              8
+              [8 [0 65.532] 9 2 10 [6 [0 14] 0 6] 0 2]
+              8
+              [ 8
+                [7 [7 [0 131.071] 9 4 0 7] 9 4 0 15]
+                8
+                [0 2]
+                9
+                2
+                10
+                [ 6
+                  [7 [0 3] 7 [0 3] 8 [0 262.138] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 6] 0 4.062] 0 2]
+                  7
+                  [0 3]
+                  8
+                  [0 2]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 7 [0 3] 8 [0 262.138] 9 2 10 [6 [7 [0 3] 1 0] [7 [0 3] 1 11] 0 4.062] 0 2]
+                    7
+                    [0 3]
+                    7
+                    [0 3]
+                    8
+                    [0 262.138]
+                    9
+                    2
+                    10
+                    [6 [7 [0 3] 1 0] [7 [0 3] 1 25] 0 4.062]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              8
+              [ 8
+                [7 [7 [0 262.143] 9 4 0 7] 9 4 0 15]
+                9
+                2
+                10
+                [ 6
+                  [7 [0 3] 8 [7 [7 [0 262.143] 9 4 0 7] 9 4 0 1] 9 2 10 [6 [0 8.158] 0 16.318] 0 2]
+                  7
+                  [0 3]
+                  8
+                  [7 [7 [0 262.143] 9 4 0 7] 9 4 0 1]
+                  9
+                  2
+                  10
+                  [6 [7 [0 3] 8 [0 1.048.567] 9 2 10 [6 0 8.158] 0 2] 0 32.638]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              8
+              [ 8
+                [0 524.284]
+                8
+                [0 2]
+                9
+                2
+                10
+                [ 6
+                  [0 130.943]
+                  7
+                  [0 3]
+                  8
+                  [0 2]
+                  9
+                  2
+                  10
+                  [ 6
+                    [0 30]
+                    7
+                    [0 3]
+                    8
+                    [0 2]
+                    9
+                    2
+                    10
+                    [ 6
+                      [0 14]
+                      7
+                      [0 3]
+                      8
+                      [0 2]
+                      9
+                      2
+                      10
+                      [ 6
+                        [7 [0 3] 7 [0 3] 8 [0 1.048.573] 9 2 10 [6 [0 1.022] 0 32.766] 0 2]
+                        7
+                        [0 3]
+                        7
+                        [0 3]
+                        8
+                        [0 1.048.573]
+                        9
+                        2
+                        10
+                        [6 [0 1.022] 0 2.046]
+                        0
+                        2
+                      ]
+                      0
+                      2
+                    ]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              9
+              2
+              10
+              [14 4 0 1.022]
+              10
+              [ 6
+                [8 [0 1.048.572] 9 2 10 [6 [0 6] 0 62] 0 2]
+                [0 1.020]
+                [0 2.042]
+                [0 4.086]
+                [8 [0 1.048.572] 9 2 10 [6 [0 16.366] 0 6] 0 2]
+                [0 16.350]
+                [0 32.702]
+                0
+                65.406
+              ]
+              0
+              127
+            ]
+            9
+            2
+            0
+            1
+          ]
+          9
+          2
+          0
+          1
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 2.036.426.867 [0 7] 0]
+      0
+      1
+    ]
+    7
+    [ 8
+      [1 0]
+      [ 1
+        [ 8
+          [1 0]
+          [ 1
+            8
+            [8 [9 23 0 7] 9 2 10 [6 0 14] 0 2]
+            [0 2]
+            10
+            [6 8 [9 10 0 63] 9 2 10 [6 [7 [0 3] 1 1.932.355.439] 7 [0 3] 8 [9 4 0 255] 9 2 10 [6 [0 126] 0 6] 0 2] 0 2]
+            0
+            15
+          ]
+          0
+          1
+        ]
+        [ [ 8
+            [1 0]
+            [ 1
+              11
+              [1.851.876.717 [1 [1 1.717.658.988] 7 [0 1] 8 [1 1 114 97 100 45 122 101 114 111 0] 9 2 0 1] 0 1]
+              6
+              [5 [1 0] 0 6]
+              [0 0]
+              8
+              [8 [9 23 0 7] 9 2 10 [6 7 [0 3] 8 [7 [7 [0 1] 9 4 0 31] 9 686 0 1] 9 2 10 [6 [7 [0 3] 1 0] 0 14] 0 2] 0 2]
+              6
+              [8 [7 [7 [0 3] 9 4 0 31] 9 343 0 255] 9 2 10 [6 [0 6] 0 30] 0 2]
+              [0 2]
+              9
+              2
+              10
+              [30 4 0 62]
+              0
+              3
+            ]
+            0
+            1
+          ]
+          [10 [6 8 [9 10 0 7] 9 2 10 [6 [7 [0 3] 1 30.449.275.492.921.459] 0 14] 0 2] 0 1]
+          10
+          [6 8 [9 10 0 7] 9 2 10 [6 [7 [0 3] 1 32.138.125.353.185.395] 0 14] 0 2]
+          0
+          1
+        ]
+        [ 8
+          [1 0]
+          [ 1
+            8
+            [8 [9 20 0 7] 9 2 10 [6 0 14] 0 2]
+            [0 2]
+            10
+            [6 8 [9 10 0 63] 9 2 10 [6 [7 [0 3] 1 1.932.355.439] 7 [0 3] 8 [9 4 0 255] 9 2 10 [6 [0 126] 0 6] 0 2] 0 2]
+            0
+            15
+          ]
+          0
+          1
+        ]
+        7
+        [ 8
+          [1 0]
+          [ 1
+            8
+            [7 [7 [0 1] 9 4 0 31] 9 46 0 1]
+            9
+            2
+            10
+            [ 6
+              [7 [0 3] 1 0]
+              7
+              [0 3]
+              8
+              [ 8
+                [7 [7 [0 1] 9 4 0 31] 9 10 0 3]
+                9
+                2
+                10
+                [6 [7 [0 3] 1 1.630.365.551] 7 [0 3] 8 [7 [7 [0 1] 9 4 0 31] 9 4 0 15] 9 2 10 [6 [0 14] 0 62] 0 2]
+                0
+                2
+              ]
+              8
+              [ 1
+                6
+                [5 [1 0] 0 30]
+                [1 0]
+                8
+                [ 8
+                  [7 [7 [0 7] 9 4 0 31] 9 10 0 3]
+                  9
+                  2
+                  10
+                  [ 6
+                    [7 [0 3] 1 1.647.142.767]
+                    7
+                    [0 3]
+                    8
+                    [7 [7 [0 7] 9 4 0 31] 9 4 0 15]
+                    9
+                    2
+                    10
+                    [6 [0 62] 7 [0 3] 8 [7 [7 [0 7] 9 4 0 31] 9 4 0 15] 9 2 10 [6 [0 254] 0 14] 0 2]
+                    0
+                    2
+                  ]
+                  0
+                  2
+                ]
+                6
+                [8 [7 [7 [0 15] 9 4 0 31] 9 343 0 255] 9 2 10 [6 [0 126] 7 [0 3] 1 256] 0 2]
+                [[[0 62] 8 [7 [7 [0 15] 9 4 0 31] 9 1.375 0 1] 9 2 10 [6 [7 [0 3] [1 0] 0 62] 0 6] 0 2] 1 0]
+                [[1 256] 0 2]
+                9
+                2
+                10
+                [30 8 [7 [7 [0 15] 9 4 0 31] 9 47 0 255] 9 2 10 [6 [0 126] 7 [0 3] 1 256] 0 2]
+                10
+                [6 0 2]
+                0
+                3
+              ]
+              9
+              2
+              0
+              1
+            ]
+            0
+            2
+          ]
+          0
+          1
+        ]
+        11
+        [1.953.718.630 1 7.823.730 [0 7] 0]
+        0
+        1
+      ]
+      0
+      1
+    ]
+    11
+    [1.953.718.630 1 26.479 [0 7] 0]
+    0
+    1
+  ]
+  [ [7 [8 [1 0 0] [1 8 [1 133.480.762.729.846] 0 0] 0 1] 11 [1.953.718.630 1 133.480.762.729.846 [0 7] 0] 0 1]
+    [7 [8 [1 0 0] [1 8 [1 1.852.270.963] 0 0] 0 1] 11 [1.953.718.630 1 1.852.270.963 [0 7] 0] 0 1]
+    [ 7
+      [8 [1 0 0 0] [1 8 [1 133.449.338.811.408.550.175.386.169.016.308.622.710] 0 0] 0 1]
+      11
+      [1.953.718.630 1 133.449.338.811.408.550.175.386.169.016.308.622.710 [0 7] 0]
+      0
+      1
+    ]
+    7
+    [8 [1 0 0] [1 8 [1 2.036.275.311.453.377.535.635.164.748.015.987] 0 0] 0 1]
+    11
+    [1.953.718.630 1 2.036.275.311.453.377.535.635.164.748.015.987 [0 7] 0]
+    0
+    1
+  ]
+  [ [ 7
+      [ 8
+        [1 0 0]
+        [ 1
+          8
+          [1 0 0]
+          8
+          [ 1
+            6
+            [6 [5 [1 0] 0 60] [6 [5 [1 0] 0 61] [1 0] 1 1] 1 1]
+            [0 13]
+            9
+            2
+            10
+            [ 30
+              [8 [8 [9 10 0 63] 9 767 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 [7 [0 3] 1 1] 0 124] 0 2]
+              8
+              [8 [9 10 0 63] 9 767 10 [6 7 [0 3] 1 0] 0 2]
+              9
+              2
+              10
+              [6 [7 [0 3] 1 1] 0 125]
+              0
+              2
+            ]
+            10
+            [ 6
+              [4 0 12]
+              8
+              [9 20 0 511]
+              9
+              2
+              10
+              [ 6
+                [0 29]
+                7
+                [0 3]
+                8
+                [8 [9 10 0 63] 9 90 10 [6 7 [0 3] 1 0] 0 2]
+                9
+                2
+                10
+                [ 6
+                  [0 28]
+                  7
+                  [0 3]
+                  5
+                  [8 [8 [9 10 0 63] 9 367 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 [7 [0 3] 1 1] 0 124] 0 2]
+                  8
+                  [8 [9 10 0 63] 9 367 10 [6 7 [0 3] 1 0] 0 2]
+                  9
+                  2
+                  10
+                  [6 [7 [0 3] 1 1] 0 125]
+                  0
+                  2
+                ]
+                0
+                2
+              ]
+              0
+              2
+            ]
+            0
+            1
+          ]
+          9
+          2
+          0
+          1
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.891.309 [0 7] 0]
+      0
+      1
+    ]
+    [ [ 8
+        [[8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1] 8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        [1 8 [[0 50] 0 54] [1 8 [[8 [0 60] 9 2 10 [6 0 28] 0 2] 8 [0 61] 9 2 10 [6 0 29] 0 2] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        0
+        1
+      ]
+      [ 8
+        [[8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1] 8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        [1 8 [9 46 0 7] 9 2 10 [6 7 [0 3] 8 [9 20 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2]
+        0
+        1
+      ]
+      7
+      [ 8
+        [1 0]
+        [ 1
+          6
+          [5 [1 0] 0 6]
+          [1 1 1]
+          8
+          [8 [8 [9 10 0 15] 9 190 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 0 14] 0 2]
+          8
+          [8 [8 [9 10 0 31] 9 190 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 0 6] 0 2]
+          [8 [9 20 0 511] 9 2 10 [6 [7 [0 3] 8 [9 20 0 511] 9 2 10 [6 [0 6] 0 6] 0 2] 0 14] 0 2]
+          8
+          [8 [9 10 0 63] 9 4 10 [6 7 [0 3] 1 0] 0 2]
+          9
+          2
+          10
+          [ 6
+            [7 [0 3] 8 [9 4 0 63] 9 2 10 [6 0 6] 0 2]
+            7
+            [0 3]
+            8
+            [9 4 0 31]
+            9
+            2
+            10
+            [ 6
+              [7 [0 3] 8 [8 [9 10 0 63] 9 367 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 [7 [0 3] 8 [9 342 0 511] 9 2 10 [6 0 6] 0 2] 0 14] 0 2]
+              7
+              [0 3]
+              8
+              [8 [9 10 0 63] 9 90 10 [6 7 [0 3] 1 0] 0 2]
+              9
+              2
+              10
+              [6 [7 [0 3] 8 [9 342 0 511] 9 2 10 [6 0 6] 0 2] 0 62]
+              0
+              2
+            ]
+            0
+            2
+          ]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.627.117 [0 7] 0]
+      0
+      1
+    ]
+    [ 7
+      [ 8
+        [1 0]
+        [ 1
+          8
+          [1 0]
+          7
+          [ 8
+            [ 1
+              6
+              [6 [3 0 30] [1 1] 1 0]
+              [ 8
+                [8 [9 43 0 31] 9 2 10 [6 0 62] 0 2]
+                [8 [9 20 0 1.023] 9 2 10 [6 [7 [0 3] 1 1] 0 12] 0 2]
+                8
+                [8 [9 10 0 127] 9 90 10 [6 7 [0 3] 1 0] 0 2]
+                9
+                2
+                10
+                [6 [7 [0 3] 1 1] 0 13]
+                0
+                2
+              ]
+              7
+              [10 [6 8 [9 20 0 511] 9 2 10 [6 [7 [0 3] 1 2] 0 14] 0 2] 0 1]
+              8
+              [9 2 10 [30 0 60] 0 1]
+              8
+              [9 2 10 [30 0 125] 10 [6 8 [9 20 0 1.023] 9 2 10 [6 [0 30] 0 12] 0 2] 0 3]
+              [8 [9 20 0 2.047] 9 2 10 [6 [7 [0 3] 1 2] 7 [0 3] 8 [9 20 0 2.047] 9 2 10 [6 [0 28] 0 12] 0 2] 0 2]
+              8
+              [9 4 0 127]
+              9
+              2
+              10
+              [ 6
+                [7 [0 3] 1 1]
+                7
+                [0 3]
+                8
+                [8 [9 10 0 255] 9 90 10 [6 7 [0 3] 1 0] 0 2]
+                9
+                2
+                10
+                [6 [7 [0 3] 1 2] 7 [0 3] 8 [8 [9 10 0 255] 9 4 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 [0 29] 0 13] 0 2]
+                0
+                2
+              ]
+              0
+              2
+            ]
+            9
+            2
+            0
+            1
+          ]
+          0
+          3
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.168.362 [0 7] 0]
+      0
+      1
+    ]
+    [ 8
+      [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+      [ 1
+        8
+        [1 0]
+        [ 1
+          8
+          [ 6
+            [3 0 6]
+            [ [8 [0 30] 9 2 10 [6 0 28] 0 2]
+              [8 [7 [0 7] 8 [9 46 0 7] 9 2 10 [6 0 14] 0 2] 9 2 10 [6 0 58] 0 2]
+              8
+              [7 [0 7] 8 [9 46 0 7] 9 2 10 [6 0 14] 0 2]
+              9
+              2
+              10
+              [6 0 59]
+              0
+              2
+            ]
+            6
+            [5 [1 0] 0 6]
+            [1 0]
+            0
+            0
+          ]
+          8
+          [5 [0 14] 0 2]
+          0
+          6
+        ]
+        0
+        1
+      ]
+      0
+      1
+    ]
+    [ 7
+      [ 8
+        [1 0]
+        [ 1
+          8
+          [1 0]
+          7
+          [ 8
+            [ 1
+              6
+              [5 [1 0] 8 [8 [9 10 0 63] 9 94 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 [7 [0 3] [0 6] 1 1] 0 62] 0 2]
+              [8 [8 [9 95 0 31] 9 2 10 [6 [7 [0 3] 4 0 6] 0 62] 0 2] [4 0 4] 0 5]
+              8
+              [8 [9 20 0 511] 9 2 10 [6 [7 [0 3] 1 2] 0 14] 0 2]
+              6
+              [5 [1 0] 8 [8 [9 10 0 127] 9 94 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 [7 [0 3] [4 0 14] 1 1] 0 126] 0 2]
+              [ 8
+                [9 2 10 [6 0 2] 0 3]
+                8
+                [9 2 10 [6 8 [9 20 0 2.047] 9 2 10 [6 [0 12] 0 14] 0 2] 0 7]
+                8
+                [[0 13] 0 5]
+                [8 [9 20 0 8.191] 9 2 10 [6 [7 [0 3] 1 2] 7 [0 3] 8 [9 20 0 8.191] 9 2 10 [6 [0 60] 0 28] 0 2] 0 2]
+                0
+                2
+              ]
+              0
+              0
+            ]
+            9
+            2
+            0
+            1
+          ]
+          0
+          3
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 6.649.187 [0 7] 0]
+      0
+      1
+    ]
+    7
+    [ 8
+      [1 0 0]
+      [ 1
+        8
+        [ 8
+          [[1 0] 8 [8 [9 10 0 15] 9 190 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 0 29] 0 2]
+          8
+          [ 1
+            6
+            [8 [9 43 0 511] 9 2 10 [6 [0 28] 0 29] 0 2]
+            [0 0]
+            6
+            [ 5
+              [1 0]
+              8
+              [8 [9 10 0 63] 9 94 10 [6 7 [0 3] 1 0] 0 2]
+              9
+              2
+              10
+              [6 [7 [0 3] [8 [9 20 0 511] 9 2 10 [6 [0 124] 0 28] 0 2] 1 1] 0 125]
+              0
+              2
+            ]
+            [9 2 10 [12 4 0 12] 0 1]
+            0
+            12
+          ]
+          9
+          2
+          0
+          1
+        ]
+        6
+        [5 [1 0] 0 2]
+        [1 1 0]
+        8
+        [8 [9 20 0 255] 9 2 10 [6 [0 60] 7 [0 3] 4 0 2] 0 2]
+        8
+        [ 8
+          [9 20 0 511]
+          9
+          2
+          10
+          [ 6
+            [7 [0 3] 8 [9 4 0 63] 9 2 10 [6 7 [0 3] 8 [9 342 0 511] 9 2 10 [6 0 14] 0 2] 0 2]
+            7
+            [0 3]
+            8
+            [8 [9 10 0 63] 9 94 10 [6 7 [0 3] 1 0] 0 2]
+            9
+            2
+            10
+            [6 [7 [0 3] [0 2] 8 [9 342 0 511] 9 2 10 [6 0 14] 0 2] 0 125]
+            0
+            2
+          ]
+          0
+          2
+        ]
+        [8 [9 20 0 1.023] 9 2 10 [6 [7 [0 3] 8 [9 20 0 1.023] 9 2 10 [6 [0 30] 0 30] 0 2] 0 6] 0 2]
+        8
+        [8 [9 10 0 127] 9 94 10 [6 7 [0 3] 1 0] 0 2]
+        9
+        2
+        10
+        [6 [7 [0 3] [8 [9 20 0 1.023] 9 2 10 [6 [0 14] 7 [0 3] 8 [9 342 0 1.023] 9 2 10 [6 0 30] 0 2] 0 2] 0 2] 0 253]
+        0
+        2
+      ]
+      0
+      1
+    ]
+    11
+    [1.953.718.630 1 6.452.594 [0 7] 0]
+    0
+    1
+  ]
+  [ [8 [1 0] [1 6 [5 [1 0] 0 6] [1 1] 8 [9 4 0 63] 9 2 10 [6 [7 [0 3] 1 2] 7 [0 3] 9 2 10 [6 8 [9 342 0 63] 9 2 10 [6 0 14] 0 2] 0 1] 0 2] 0 1]
+    [ 8
+      [1 0]
+      [ 1
+        [ 8
+          [1 0 0]
+          [1 8 [9 20 0 255] 9 2 10 [6 [7 [0 3] 8 [9 90 0 7] 9 2 10 [6 [7 [0 3] 8 [9 190 0 7] 9 2 10 [6 0 28] 0 2] 0 29] 0 2] 0 28] 0 2]
+          0
+          1
+        ]
+        [8 [1 0] [1 8 [9 367 0 7] 9 2 10 [6 [7 [0 3] 1 1] 0 14] 0 2] 0 1]
+        [ [ 8
+            [1 0 0]
+            [ 1
+              8
+              [1 0]
+              7
+              [10 [29 8 [9 10 0 15] 9 2 10 [6 0 61] 0 2] 0 1]
+              8
+              [0 29]
+              8
+              [ 1
+                6
+                [5 [0 14] 0 124]
+                [8 [9 767 0 63] 9 2 10 [6 [7 [0 3] 1 1] 0 14] 0 2]
+                9
+                2
+                10
+                [14 4 0 14]
+                10
+                [6 8 [9 20 0 2.047] 9 2 10 [6 [0 253] 7 [0 3] 8 [9 90 0 63] 9 2 10 [6 [7 [0 3] 1 1] 0 14] 0 2] 0 2]
+                0
+                1
+              ]
+              9
+              2
+              0
+              1
+            ]
+            0
+            1
+          ]
+          [ 8
+            [1 0 0]
+            [ 1
+              8
+              [9 4 0 255]
+              9
+              2
+              10
+              [6 [0 29] 7 [0 3] 8 [9 4 0 31] 9 2 10 [6 7 [0 3] 8 [9 4 0 255] 9 2 10 [6 [7 [0 3] 9 182 0 7] 0 28] 0 2] 0 2]
+              0
+              2
+            ]
+            0
+            1
+          ]
+          [8 [9 4 0 7] 9 2 10 [6 0 14] 0 2]
+          [ 8
+            [1 0]
+            [1 8 [9 47 0 255] 9 2 10 [6 [7 [0 3] 8 [9 342 0 255] 9 2 10 [6 7 [0 3] 9 382 0 7] 0 2] 7 [0 3] 8 [9 10 0 7] 9 2 10 [6 0 14] 0 2] 0 2]
+            0
+            1
+          ]
+          8
+          [1 0 0]
+          [ 1
+            8
+            [9 46 0 255]
+            9
+            2
+            10
+            [6 [0 29] 7 [0 3] 8 [9 4 0 31] 9 2 10 [6 7 [0 3] 8 [9 4 0 255] 9 2 10 [6 [7 [0 3] 9 182 0 7] 0 28] 0 2] 0 2]
+            0
+            2
+          ]
+          0
+          1
+        ]
+        [8 [1 0 0] [1 8 [9 10 0 7] 9 2 10 [6 7 [0 3] 8 [9 20 0 255] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2] 0 1]
+        [8 [1 [0 0] 0] [1 8 [9 367 0 7] 9 2 10 [6 [0 57] 7 [0 3] 8 [9 767 0 7] 9 2 10 [6 [0 56] 0 29] 0 2] 0 2] 0 1]
+        [ 8
+          [1 0]
+          [1 8 [1 0] 8 [1 6 [5 [1 0] 0 30] [0 6] 9 2 10 [30 8 [9 767 0 31] 9 2 10 [6 [7 [0 3] 1 1] 0 62] 0 2] 10 [6 4 0 6] 0 1] 9 2 0 1]
+          0
+          1
+        ]
+        [8 [9 4 0 7] 9 2 10 [6 7 [0 3] 8 [9 4 0 7] 9 2 10 [6 0 14] 0 2] 0 2]
+        [8 [1 0] [1 8 [9 46 0 7] 9 2 10 [6 [7 [0 3] 8 [9 366 0 7] 9 2 10 [6 0 14] 0 2] 7 [0 3] 1 1] 0 2] 0 1]
+        8
+        [1 0 0]
+        [ 1
+          8
+          [9 170 0 255]
+          9
+          2
+          10
+          [6 [0 29] 7 [0 3] 8 [9 4 0 31] 9 2 10 [6 7 [0 3] 8 [9 4 0 255] 9 2 10 [6 [7 [0 3] 9 182 0 7] 0 28] 0 2] 0 2]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      0
+      1
+    ]
+    8
+    [1 0]
+    [1 8 [8 [9 10 0 7] 9 190 10 [6 7 [0 3] 1 0] 0 2] 9 2 10 [6 0 14] 0 2]
+    0
+    1
+  ]
+  [ [ 7
+      [8 [1 0 0] [1 6 [5 [1 0] 0 13] [1 1] 8 [9 4 0 31] 9 2 10 [6 [0 28] 7 [0 3] 9 2 10 [13 8 [9 342 0 31] 9 2 10 [6 0 29] 0 2] 0 1] 0 2] 0 1]
+      11
+      [1.953.718.630 1 7.827.312 [0 7] 0]
+      0
+      1
+    ]
+    8
+    [1 0]
+    [ 1
+      [8 [1 1 1] [1 8 [9 42 0 7] 9 2 10 [6 7 [0 3] 8 [9 4 0 127] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2] 0 1]
+      [ [8 [1 0 0] [1 8 [9 42 0 7] 9 2 10 [6 7 [0 3] 8 [9 20 0 127] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2] 0 1]
+        [8 [1 0] [1 8 [9 46 0 127] 9 2 10 [6 [0 14] 0 62] 0 2] 0 1]
+        [8 [1 1 1] [1 8 [9 4 0 7] 9 2 10 [6 [0 28] 7 [0 3] 8 [9 174 0 7] 9 2 10 [6 0 29] 0 2] 0 2] 0 1]
+        [8 [1 0] [1 0 0] 0 1]
+        8
+        [1 0]
+        [1 8 [9 47 0 127] 9 2 10 [6 [0 62] 7 [0 3] 8 [9 42 0 7] 9 2 10 [6 0 14] 0 2] 0 2]
+        0
+        1
+      ]
+      [8 [1 0 0] [1 5 [8 [9 42 0 7] 9 2 10 [6 0 28] 0 2] 8 [9 42 0 7] 9 2 10 [6 0 29] 0 2] 0 1]
+      [8 [1 0 0] [1 8 [9 42 0 7] 9 2 10 [6 7 [0 3] 8 [9 4 0 31] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2] 0 1]
+      8
+      [1 0 0]
+      [ 1
+        8
+        [9 42 0 7]
+        9
+        2
+        10
+        [6 7 [0 3] 8 [9 47 0 127] 9 2 10 [6 [7 [0 3] 8 [9 20 0 127] 9 2 10 [6 [0 62] 0 28] 0 2] 7 [0 3] 8 [9 42 0 7] 9 2 10 [6 0 29] 0 2] 0 2]
+        0
+        2
+      ]
+      0
+      1
+    ]
+    0
+    1
+  ]
+  [ [8 [1 [[0 15] [0 0] [0 0] 0] 0] [1 8 [0 101] [1 8 [0 60] 9 2 10 [6 [0 125] 0 14] 0 2] 0 1] 0 1]
+    [ [8 [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1] [1 8 [1 0] [1 8 [7 [0 7] [1 7 [0 14] 9 2 0 1] 0 1] 8 [5 [0 14] 0 2] 0 6] 0 1] 0 1]
+      [ 8
+        [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        [ 1
+          8
+          [1 0]
+          [1 8 [6 [3 0 6] [[6 [5 [1 0] 0 12] [1 0] 0 0] 8 [0 30] 9 2 10 [6 0 29] 0 2] 6 [5 [1 0] 0 6] [1 0] 0 0] 8 [5 [0 14] 0 2] 0 6]
+          0
+          1
+        ]
+        0
+        1
+      ]
+      [ 8
+        [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        [ 1
+          8
+          [[0 26] 7 [8 [9 47 0 7] 9 2 10 [6 0 14] 0 2] 0 6]
+          [1 8 [[8 [0 30] 9 2 10 [6 0 28] 0 2] 8 [7 [0 7] 8 [9 47 0 7] 9 2 10 [6 0 14] 0 2] 9 2 10 [6 0 29] 0 2] 8 [5 [0 14] 0 2] 0 6]
+          0
+          1
+        ]
+        0
+        1
+      ]
+      [8 [1 0] [1 8 [7 [1 0 0] 8 [0 2] [1 0 15] 0 1] 8 [5 [0 14] 0 2] 0 6] 0 1]
+      [ 8
+        [[8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1] 8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        [ 1
+          8
+          [[1 0] 0 54]
+          [ 1
+            8
+            [ 6
+              [5 [1 1] 0 12]
+              [[6 [5 [0 12] 1 1] [1 1] 0 0] 8 [0 60] 9 2 10 [6 0 29] 0 2]
+              [6 [5 [0 12] 1 0] [1 0] 0 0]
+              8
+              [0 61]
+              9
+              2
+              10
+              [6 0 29]
+              0
+              2
+            ]
+            8
+            [5 [0 14] 0 2]
+            0
+            6
+          ]
+          0
+          1
+        ]
+        0
+        1
+      ]
+      [ 8
+        [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+        [ 1
+          8
+          [1 0]
+          [ 1
+            8
+            [ 6
+              [3 0 6]
+              [ [8 [0 30] 9 2 10 [6 0 28] 0 2]
+                [8 [7 [0 7] 8 [9 702 0 7] 9 2 10 [6 0 14] 0 2] 9 2 10 [6 0 58] 0 2]
+                8
+                [7 [0 7] 8 [9 702 0 7] 9 2 10 [6 0 14] 0 2]
+                9
+                2
+                10
+                [6 0 59]
+                0
+                2
+              ]
+              6
+              [5 [1 0] 0 6]
+              [1 0]
+              0
+              0
+            ]
+            8
+            [5 [0 14] 0 2]
+            0
+            6
+          ]
+          0
+          1
+        ]
+        0
+        1
+      ]
+      [8 [1 0] [1 6 [5 [1 0] 0 6] [1 0] 4 9 2 10 [6 0 13] 0 1] 0 1]
+      8
+      [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+      [1 8 [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1] 8 [5 [0 14] 0 2] 0 6]
+      0
+      1
+    ]
+    [8 [1 0 [0 15] 0 0 0] [1 8 [1 6 [5 [1 0] 0 28] [1 0] [8 [0 29] 9 2 10 [6 0 120] 0 2] 9 2 10 [28 0 57] 0 1] 9 2 0 1] 0 1]
+    [8 [1 0 [0 13] [0 0] 0] [1 8 [1 6 [5 [1 0] 0 28] [0 237] 8 [0 29] 9 2 10 [6 [0 120] 7 [0 3] 9 2 10 [28 0 57] 0 1] 0 2] 9 2 0 1] 0 1]
+    8
+    [8 [1 0] [1 8 [0 6] 8 [5 [0 14] 0 2] 0 6] 0 1]
+    [ 1
+      8
+      [1 0]
+      [ 1
+        8
+        [6 [3 0 6] [[8 [0 30] 9 2 10 [6 0 28] 0 2] 8 [7 [0 7] 8 [9 47 0 7] 9 2 10 [6 0 14] 0 2] 9 2 10 [6 0 29] 0 2] 6 [5 [1 0] 0 6] [1 0] 0 0]
+        8
+        [5 [0 14] 0 2]
+        0
+        6
+      ]
+      0
+      1
+    ]
+    0
+    1
+  ]
+  [ [ 7
+      [ 8
+        [1 1 1]
+        [ 1
+          8
+          [1 0]
+          8
+          [1 6 [5 [1 0] 0 60] [0 6] 9 2 10 [60 8 [9 342 0 31] 9 2 10 [6 0 124] 0 2] 10 [6 8 [9 20 0 31] 9 2 10 [6 [0 125] 0 14] 0 2] 0 1]
+          9
+          2
+          0
+          1
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 7.107.949 [0 7] 0]
+      0
+      1
+    ]
+    [ [ 7
+        [8 [1 0 0] [1 6 [5 [1 0] 0 12] [0 13] 9 2 10 [6 [8 [9 342 0 7] 9 2 10 [6 0 28] 0 2] 4 0 13] 0 1] 0 1]
+        11
+        [1.953.718.630 1 6.579.297 [0 7] 0]
+        0
+        1
+      ]
+      [ [ 7
+          [8 [1 0 0] [1 6 [5 [0 12] 0 13] [1 0] 6 [8 [9 343 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] [1 0] 1 1] 0 1]
+          11
+          [1.953.718.630 1 6.648.940 [0 7] 0]
+          0
+          1
+        ]
+        [ 7
+          [ 8
+            [1 1 1]
+            [ 1
+              6
+              [5 [1 0] 0 13]
+              [0 0]
+              8
+              [1 0]
+              8
+              [1 6 [8 [9 343 0 31] 9 2 10 [6 [0 124] 0 125] 0 2] [0 6] 9 2 10 [60 8 [9 47 0 31] 9 2 10 [6 [0 124] 0 125] 0 2] 10 [6 4 0 6] 0 1]
+              9
+              2
+              0
+              1
+            ]
+            0
+            1
+          ]
+          11
+          [1.953.718.630 1 7.760.228 [0 7] 0]
+          0
+          1
+        ]
+        [ 7
+          [8 [1 0] [1 6 [5 [1 0] 0 6] [0 0] 8 [1 0] 8 [1 6 [5 [0 30] 4 0 6] [0 6] 9 2 10 [6 4 0 6] 0 1] 9 2 0 1] 0 1]
+          11
+          [1.953.718.630 1 6.514.020 [0 7] 0]
+          0
+          1
+        ]
+        7
+        [ 8
+          [1 0 0]
+          [ 1
+            6
+            [6 [5 [0 12] 0 13] [1 1] 1 0]
+            [ 6
+              [ 8
+                [ 1
+                  6
+                  [5 [1 0] 0 28]
+                  [1 0]
+                  6
+                  [ 6
+                    [6 [5 [1 0] 0 29] [1 1] 1 0]
+                    [6 [9 2 10 [14 [8 [9 342 0 15] 9 2 10 [6 0 60] 0 2] 8 [9 342 0 15] 9 2 10 [6 0 61] 0 2] 0 1] [1 0] 1 1]
+                    1
+                    1
+                  ]
+                  [1 0]
+                  1
+                  1
+                ]
+                9
+                2
+                0
+                1
+              ]
+              [1 0]
+              1
+              1
+            ]
+            1
+            1
+          ]
+          0
+          1
+        ]
+        11
+        [1.953.718.630 1 6.845.548 [0 7] 0]
+        0
+        1
+      ]
+      7
+      [8 [1 0 0] [1 6 [8 [9 84 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] [1 1] 1 0] 0 1]
+      11
+      [1.953.718.630 1 6.845.543 [0 7] 0]
+      0
+      1
+    ]
+    [7 [8 [1 0 0] [1 6 [8 [9 343 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] [1 1] 1 0] 0 1] 11 [1.953.718.630 1 6.648.935 [0 7] 0] 0 1]
+    [ 7
+      [ 8
+        [1 1 1]
+        [ 1
+          6
+          [5 [1 0] 0 13]
+          [0 0]
+          8
+          [9 47 0 7]
+          9
+          2
+          10
+          [6 [0 28] 7 [0 3] 8 [9 4 0 7] 9 2 10 [6 [0 29] 7 [0 3] 8 [9 170 0 7] 9 2 10 [6 [0 28] 0 29] 0 2] 0 2]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      11
+      [1.953.718.630 1 6.582.125 [0 7] 0]
+      0
+      1
+    ]
+    7
+    [8 [1 0 0] [1 6 [5 [1 0] 0 13] [0 12] 9 2 10 [6 [8 [9 342 0 7] 9 2 10 [6 0 28] 0 2] 8 [9 342 0 7] 9 2 10 [6 0 29] 0 2] 0 1] 0 1]
+    11
+    [1.953.718.630 1 6.452.595 [0 7] 0]
+    0
+    1
+  ]
+  [0 3]
+  909
+  ]
+  """
+
+  # evaluated at compile time.
+  @stdlib_core_val Noun.Format.parse_always(stdlib_string)
+
+  @spec stdlib_core :: Noun.t()
+  def stdlib_core do
+    @stdlib_core_val
+  end
+
+  rm_string = """
+  [ [7 [8 [1 0 0 0 0] [1 8 [1 30.160.793.233.665.617.219.451.904.865] 0 0] 0 1] 11 [1.953.718.630 1 30.160.793.233.665.617.219.451.904.865 [0 7] 0] 0 1]
+  [ [1 0]
+    [ [ 8
+        [1 0 0 0]
+        [ 1
+          8
+          [ [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 751 0 1] 0 2] 9 2 10 [6 0 28] 0 2]
+            [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 751 0 1] 0 2] 9 2 10 [6 0 58] 0 2]
+            0
+            27
+          ]
+          8
+          [5 [0 14] 0 2]
+          0
+          6
+        ]
+        0
+        1
+      ]
+      [8 [1 0 [1 0] 0 0] [1 8 [[6 [6 [3 0 12] [1 1] 1 0] [0 12] 0 0] 8 [7 [0 7] 9 1.534 0 1] 9 2 10 [6 0 29] 0 2] 8 [5 [0 14] 0 2] 0 6] 0 1]
+      [8 [1 [0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102] [0 0 0 0] 0 0 0] [1 8 [9 22 0 127] 9 2 10 [6 0 14] 0 2] 0 1]
+      8
+      [1 0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102]
+      [1 [[8 [9 1.492 0 7] 9 2 10 [6 0 14] 0 2] 8 [9 10 0 15] 9 2 10 [6 0 238] 0 2] 1 0]
+      0
+      1
+    ]
+    [8 [1 0] [1 8 [6 [6 [3 0 6] [1 1] 1 0] [0 6] 0 0] 8 [5 [0 14] 0 2] 0 6] 0 1]
+    [ 8
+      [1 478.793.196.187.462.788.804.451]
+      [1 8 [6 [5 [0 6] 1 478.793.196.187.462.788.804.451] [1 478.793.196.187.462.788.804.451] 0 0] 8 [5 [0 14] 0 2] 0 6]
+      0
+      1
+    ]
+    [ 8
+      [1 0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102]
+      [1 8 [8 [9 10 0 255] 9 4 10 [6 7 [0 3] 1 3] 0 2] 9 2 10 [6 [7 [0 3] 1 6.243.918] 7 [0 3] 8 [9 22 0 127] 9 2 10 [6 0 14] 0 2] 0 2]
+      0
+      1
+    ]
+    8
+    [1 0]
+    [1 8 [6 [6 [3 0 6] [1 1] 1 0] [0 6] 0 0] 8 [5 [0 14] 0 2] 0 6]
+    0
+    1
+  ]
+  [8 [1 0] [1 1 478.793.196.187.462.788.804.451] 0 1]
+  [ [7 [8 [1 0 0] [1 8 [1 1.851.907.519.744.077.227.364] 0 0] 0 1] 11 [1.953.718.630 1 1.851.907.519.744.077.227.364 [0 7] 0] 0 1]
+    [ [8 [1 0] [1 5 [1 6.243.918] 8 [8 [9 10 0 255] 9 367 10 [6 7 [0 3] 1 3] 0 2] 9 2 10 [6 [7 [0 3] 1 3] 0 14] 0 2] 0 1]
+      [ [8 [1 0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102] [1 [0 12] 0 26] 0 1]
+        [ 8
+          [1 [0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102] [0 0 0 0] 0 0 0]
+          [ 1
+            8
+            [6 [5 [1 478.793.196.187.462.788.804.451] 0 6] [8 [7 [0 7] 9 174 0 1] 9 2 10 [6 0 14] 0 2] 8 [7 [0 7] 9 762 0 1] 9 2 10 [6 0 14] 0 2]
+            8
+            [5 [0 14] 0 2]
+            0
+            6
+          ]
+          0
+          1
+        ]
+        8
+        [1 0 0 0 0]
+        [ 1
+          8
+          [ [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 351 0 1] 0 2] 9 2 10 [6 0 28] 0 2]
+            [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 374 0 1] 0 2] 9 2 10 [6 0 58] 0 2]
+            [6 [6 [3 0 54] [1 1] 1 0] [0 54] 0 0]
+            0
+            55
+          ]
+          8
+          [5 [0 14] 0 2]
+          0
+          6
+        ]
+        0
+        1
+      ]
+      [7 [8 [1 0] [1 8 [1 460.217.181.910.180.775.551.341] 0 0] 0 1] 11 [1.953.718.630 1 460.217.181.910.180.775.551.341 [0 7] 0] 0 1]
+      8
+      [1 0 0 0 0]
+      [ 1
+        8
+        [ [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 351 0 1] 0 2] 9 2 10 [6 0 28] 0 2]
+          [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 374 0 1] 0 2] 9 2 10 [6 0 58] 0 2]
+          [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 2.986 0 1] 0 2] 9 2 10 [6 0 118] 0 2]
+          0
+          55
+        ]
+        8
+        [5 [0 14] 0 2]
+        0
+        6
+      ]
+      0
+      1
+    ]
+    [8 [1 0] [1 8 [6 [6 [3 0 6] [1 1] 1 0] [0 6] 0 0] 8 [5 [0 14] 0 2] 0 6] 0 1]
+    [8 [1 [0 [1 0] 0 0] 0] [1 8 [[8 [7 [0 7] 9 170 0 1] 9 2 10 [6 0 28] 0 2] 6 [6 [3 0 13] [1 1] 1 0] [0 13] 0 0] 8 [5 [0 14] 0 2] 0 6] 0 1]
+    8
+    [1 0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102]
+    [ 1
+      8
+      [ [6 [6 [3 0 12] [1 1] 1 0] [0 12] 0 0]
+        [8 [7 [0 7] 9 1.534 0 1] 9 2 10 [6 0 58] 0 2]
+        [6 [5 [1 0] 0 54] [1 0] 6 [5 [1 1] 0 54] [1 1] 0 0]
+        [6 [6 [3 0 110] [1 1] 1 0] [0 110] 0 0]
+        [[6 [6 [3 0 444] [1 1] 1 0] [0 444] 0 0] 6 [6 [3 0 445] [1 1] 1 0] [0 445] 0 0]
+        [6 [6 [3 0 446] [1 1] 1 0] [0 446] 0 0]
+        [6 [6 [3 0 894] [1 1] 1 0] [0 894] 0 0]
+        6
+        [5 [0 895] 1 1.701.536.102]
+        [1 1.701.536.102]
+        0
+        0
+      ]
+      8
+      [5 [0 14] 0 2]
+      0
+      6
+    ]
+    0
+    1
+  ]
+  [ 8
+    [1 0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102]
+    [1 8 [8 [9 10 0 255] 9 4 10 [6 7 [0 3] 1 3] 0 2] 9 2 10 [6 [7 [0 3] 1 6.245.699] 7 [0 3] 8 [9 22 0 127] 9 2 10 [6 0 14] 0 2] 0 2]
+    0
+    1
+  ]
+  [ [ [ 1
+        [ 8
+          [1 0]
+          [ 1
+            8
+            [9 751 0 15]
+            9
+            2
+            10
+            [6 7 [0 3] 8 [9 94 0 255] 9 2 10 [6 7 [0 3] 8 [8 [9 10 0 511] 9 767 10 [6 7 [0 3] 1 3] 0 2] 9 2 10 [6 [7 [0 3] 1 3] 0 14] 0 2] 0 2]
+            0
+            2
+          ]
+          0
+          1
+        ]
+        8
+        [1 0]
+        [ 1
+          8
+          [9 751 0 15]
+          9
+          2
+          10
+          [6 7 [0 3] 8 [9 94 0 255] 9 2 10 [6 7 [0 3] 8 [8 [9 10 0 511] 9 767 10 [6 7 [0 3] 1 3] 0 2] 9 2 10 [6 [7 [0 3] 1 3] 0 14] 0 2] 0 2]
+          0
+          2
+        ]
+        0
+        1
+      ]
+      0
+      1
+    ]
+    [ 8
+      [1 [0 [[1 0] 0 0] 1 1 [0 0] 0 0 1.701.536.102] [0 0 0 0] 0 0 0]
+      [ 1
+        8
+        [[8 [7 [0 7] 9 751 0 1] 9 2 10 [6 0 28] 0 2] [8 [7 [0 7] 9 2.987 0 1] 9 2 10 [6 0 58] 0 2] 8 [7 [0 7] 9 84 0 1] 9 2 10 [6 0 59] 0 2]
+        8
+        [5 [0 14] 0 2]
+        0
+        6
+      ]
+      0
+      1
+    ]
+    [8 [1 0] [1 5 [1 6.245.699] 8 [8 [9 10 0 255] 9 367 10 [6 7 [0 3] 1 3] 0 2] 9 2 10 [6 [7 [0 3] 1 3] 0 14] 0 2] 0 1]
+    7
+    [8 [1 0 0] [1 8 [1 1.816.244.077.244.883.690.852] 0 0] 0 1]
+    11
+    [1.953.718.630 1 1.816.244.077.244.883.690.852 [0 7] 0]
+    0
+    1
+  ]
+  [ 8
+    [1 0 0 0 418.565.088.612]
+    [ 1
+      8
+      [ [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 86 0 1] 0 2] 9 2 10 [6 0 28] 0 2]
+        [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 1.495 0 1] 0 2] 9 2 10 [6 0 58] 0 2]
+        [8 [7 [0 7] 9 766 0 1] 9 2 10 [6 0 118] 0 2]
+        6
+        [5 [0 55] 1 418.565.088.612]
+        [1 418.565.088.612]
+        0
+        0
+      ]
+      8
+      [5 [0 14] 0 2]
+      0
+      6
+    ]
+    0
+    1
+  ]
+  [8 [1 0] [1 8 [8 [7 [0 7] 8 [9 47 0 255] 9 2 10 [6 7 [0 3] 9 750 0 1] 0 2] 9 2 10 [6 0 14] 0 2] 8 [5 [0 14] 0 2] 0 6] 0 1]
+  [8 [1 [1 0] 0 0] [1 8 [7 [1 0] 8 [1 0] [1 1 0] 0 1] 8 [5 [0 14] 0 2] 0 6] 0 1]
+  8
+  [1 0]
+  [1 1 418.565.088.612]
+  0
+  1
+  ]
+  """
+
+  # evaluated at compile time.
+  @rm_core_val [Noun.Format.parse_always(rm_string) | @stdlib_core_val]
+
+  @spec rm_core :: Noun.t()
+  def rm_core do
+    @rm_core_val
+  end
+
+  # temporarily stubbed out
+  logics_string = """
+  [123 456]
+  """
+
+  # evaluated at compile time.
+  @logics_core_val [Noun.Format.parse_always(logics_string) | @rm_core_val]
+
+  @spec logics_core :: Noun.t()
+  def logics_core do
+    @logics_core_val
+  end
+end
